Torch-Flask-ImageCartoonization/Cartoonization API/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu
Werkzeug==2.0.1
cloudinary==1.26.0
Flask==1.1.2
gunicorn==20.1.0
opencv-python-headless==4.5.3.56
tqdm==4.59.0
requests==2.25.1
numpy==1.20.2
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
Pillow==8.3.2
python-dotenv==0.19.0
